requirements.txt
streamlit
langchain
beautifulsoup4
chromadb
python-dotenv
langchain-openai